Create a separate module for configuration.
