+++ src/nimbuscli/config/__init__.py
@@ -1,0 +1,1 @@
+from nimbuscli.config.config import Config, load_config, resolve_config
